+++ src/Modules/Settings/components/Passport.vue
@@ -41,7 +41,8 @@ export default {
     async tryCreateAsUser () {
       showLoader()
       try {
-        const blob = await createPassportAsUser()
+        const jsonData = await createPassportAsUser()
+        const blob = new Blob(jsonData.response, { type: 'application/json' })
         const filename = 'fs_passport_' + this.userDetails.id + '.pdf'
         this.downloadFile(blob, filename)
       } catch (e) {