+++ tests/Acceptance/IdCardsCest.php
@@ -21,18 +21,34 @@ class IdCardsCest
     public function AmbassadorCanCreateIdCard(AcceptanceTester $I): void
     {
         $region = $I->createRegion();
-        $I->createFoodsaver(null, ['name' => 'fs1', 'nachname' => 'saver1', 'photo' => 'does-not-exist.jpg', 'bezirk_id' => $region['id']]);
+        $foodSaver = $I->createFoodsaver(null, ['name' => 'fs1', 'nachname' => 'saver1', 'photo' => 'does-not-exist.jpg', 'bezirk_id' => $region['id']]);
         $ambassador = $I->createAmbassador(null, ['photo' => 'does-not-exist.jpg', 'bezirk_id' => $region['id']]);
         $I->addRegionAdmin($region['id'], $ambassador['id']);
 
         $I->login($ambassador['email']);
 
-        $I->amOnPage('/?page=passgen&bid=' . $region['id']);
-        $I->waitForText('fs1 saver1');
-        $I->click('Alle markieren');
-        $I->click('Markierte Ausweise generieren');
+        $I->amOnPage('/region?bid=' . $region['id'] . '&sub=members');
+        $I->waitForText('Foodsaver:innen im Bezirk');
+        $I->click('Ausweise');
+        $I->waitForElementVisible('#filterMember', 30);
+        $I->fillField('#filterMember', $foodSaver['name']);
 
-        /* ToDo: Not supported in new ci run style */
-        //$I->waitForFileExists('/downloads/foodsaver_pass_' . convertRegionName($region['name']) . '.pdf', 10);
+        $searchContent = $foodSaver['name'];
+        $I->see($searchContent);
+
+        $I->executeJS('
+            var rows = document.querySelectorAll("tr[role=\'row\']");
+            for (var i = 0; i < rows.length; i++) {
+                if (rows[i].innerText.includes("' . $searchContent . '")) {
+                    rows[i].setAttribute("aria-selected", "true");
+                } else {
+                    rows[i].setAttribute("aria-selected", "false");
+                }
+            }
+        ');
+
+        $I->click('Ausweis/e für markierte erstellen');
+
+        // $I->waitForFileExists('/downloads/fs_passports_' . $region['id'] . '_' . convertRegionName($region['name']) . '.pdf', 10);
     }
 }